implement authentation

+++ MusicEventSystem/src/main/java/com/example/musiceventsystem/controllers/CustomerServlet.java
@@ -1,5 +1,6 @@
 package com.example.musiceventsystem.controllers;
 
+import com.example.musiceventsystem.auth.Authorization;
 import com.example.musiceventsystem.model.Customer;
 import com.example.musiceventsystem.service.CustomerService;
 import jakarta.servlet.ServletException;
@@ -7,6 +8,7 @@ import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
 
 import java.io.IOException;
 
@@ -14,6 +16,7 @@ import java.io.IOException;
 public class CustomerServlet extends HttpServlet {
 
     private CustomerService customerService = new CustomerService();
+    private Authorization authorization = new Authorization();
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
@@ -22,20 +25,34 @@ public class CustomerServlet extends HttpServlet {
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        HttpSession session = req.getSession();
+        String userRole = (String) session.getAttribute("roleType");
         req.setCharacterEncoding("UTF-8");
         String method = req.getParameter("method");
         switch (method){
             case "list":
+                if (!authorization.checkPermission(userRole, "customer list")) {
+                    resp.sendRedirect("/accessdenied.jsp");
+                    return;
+                }
                 req.setAttribute("list", this.customerService.list());
                 req.getRequestDispatcher("managecustomer.jsp").forward(req, resp);
                 break;
             case "search":
+                if (!authorization.checkPermission(userRole, "customer search")) {
+                    resp.sendRedirect("/accessdenied.jsp");
+                    return;
+                }
                 String key = req.getParameter("key");
                 String value = req.getParameter("value");
                 req.setAttribute("list", this.customerService.search(key, value));
                 req.getRequestDispatcher("managecustomer.jsp").forward(req, resp);
                 break;
             case "save":
+                if (!authorization.checkPermission(userRole, "customer save")) {
+                    resp.sendRedirect("/accessdenied.jsp");
+                    return;
+                }
                 String username = req.getParameter("username");
                 String password = req.getParameter("password");
                 String name = req.getParameter("name");
@@ -44,6 +61,10 @@ public class CustomerServlet extends HttpServlet {
                 resp.sendRedirect("/customer?method=list");
                 break;
             case "update":
+                if (!authorization.checkPermission(userRole, "customer update")) {
+                    resp.sendRedirect("/accessdenied.jsp");
+                    return;
+                }
                 String idStr = req.getParameter("id");
                 Integer id = Integer.parseInt(idStr);
                 username = req.getParameter("username");
@@ -54,6 +75,10 @@ public class CustomerServlet extends HttpServlet {
                 resp.sendRedirect("/customer?method=list");
                 break;
             case "delete":
+                if (!authorization.checkPermission(userRole, "customer delete")) {
+                    resp.sendRedirect("/accessdenied.jsp");
+                    return;
+                }
                 idStr = req.getParameter("id");
                 id = Integer.parseInt(idStr);
                 this.customerService.delete(id);
